% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/denoise.R
\name{closure}
\alias{closure}
\title{Close rows to unit sum}
\usage{
closure(A)
}
\arguments{
\item{A}{nonnegative matrix}
}
\value{
\code{A} with every row divided by its sum
}
\description{
Close rows to unit sum
}
