% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logisticcf.R
\name{cv.logisticcfR}
\alias{cv.logisticcfR}
\title{Cross validation to select optimal rank of logisticcf}
\usage{
cv.logisticcfR(X, Z, max_K = 10, lambdas = c(0.01, 0.1, 1), ncores = 1)
}
\arguments{
\item{X}{observed binary matrix}

\item{Z}{binary mask indicating whether an entry should be considered when fitting logisticcf}

\item{max_K}{maximum number of ranks to try for matrix factorization}

\item{lambdas}{ridge penalty parameters}

\item{ncores}{number of cores for parallel computing, default 1}
}
\value{
optimal choice of factorizedmatrix A and B, denoised probabilities, selected rank
}
\description{
Cross validation to select optimal rank of logisticcf
}
