% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svt_utils.R
\name{log1exp}
\alias{log1exp}
\title{log one plus exponential}
\usage{
log1exp(x)
}
\arguments{
\item{x}{a vector}
}
\value{
log(1+exp(x))
}
\description{
log one plus exponential
}
\details{
This function is used to avoid numerical issues when x is too large
}
