% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logisticcf.R
\name{logisticcfR}
\alias{logisticcfR}
\title{Logistic collaborative filtering}
\usage{
logisticcfR(X, Z, K, lambda, max_iter = 1000, tol = 1e-05)
}
\arguments{
\item{X}{observed binary matrix}

\item{Z}{binary mask indicating whether an entry should be considered when fitting logisticcf}

\item{K}{the maximum number of ranks of factorized matrices}

\item{lambda}{ridge penalty parameter}

\item{max_iter}{maximum number of iterations}

\item{tol}{convergence limit of the loss}
}
\value{
a list containting the factorized functions and the  denoised probabilities
}
\description{
Logistic collaborative filtering
}
