% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svt_utils.R
\name{mask_prevalence}
\alias{mask_prevalence}
\title{Prevalence of each mask over the training entries}
\usage{
mask_prevalence(Y, train = NULL)
}
\arguments{
\item{Y}{list of T binary matrices}

\item{train}{\code{NULL} or a validated training mask, see \code{\link[=check_train]{check_train()}}}
}
\value{
numeric vector of length T
}
\description{
Prevalence of each mask over the training entries
}
\keyword{internal}
