% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmfsvt.R
\name{offset_head}
\alias{offset_head}
\title{Offset of the first mask}
\usage{
offset_head(mu)
}
\arguments{
\item{mu}{offsets, a vector of length T or a T x P matrix}
}
\value{
\code{mu[1]}, or the first row of \code{mu} for column offsets
}
\description{
Offset of the first mask
}
\keyword{internal}
