% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logisticcf.R
\name{penalized_loss}
\alias{penalized_loss}
\title{Cross entropy loss plus ridge penalty}
\usage{
penalized_loss(X, Z, A, B, lambda)
}
\arguments{
\item{X}{observed binary matrix}

\item{Z}{binary mask indicating whether an entry is included in the loss}

\item{A}{factor matrix A}

\item{B}{factor matrix B expit(A^T*B) is the denoised probability matrix}

\item{lambda}{ridge penalty parameter}
}
\value{
value of loss
}
\description{
Cross entropy loss plus ridge penalty
}
