% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svt_utils.R
\name{sigmoid}
\alias{sigmoid}
\title{Sigmoid with guard against overflow}
\usage{
sigmoid(x)
}
\arguments{
\item{x}{a numeric vector or matrix}
}
\value{
1/(1+exp(-x)), evaluated by \code{\link[stats:Logistic]{stats::plogis()}}
}
\description{
Sigmoid with guard against overflow
}
\keyword{internal}
