% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svt_utils.R
\name{spectral_norm}
\alias{spectral_norm}
\title{Largest singular value of a matrix}
\usage{
spectral_norm(M)
}
\arguments{
\item{M}{a matrix}
}
\value{
the spectral norm
}
\description{
Largest singular value of a matrix
}
\keyword{internal}
