% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logisticcf.R
\name{train_lcf}
\alias{train_lcf}
\title{Train_LCF}
\usage{
train_lcf(
  train_mask,
  validation_mask,
  observation,
  K,
  lambda,
  max_iter = 1000,
  tol = 1e-05,
  cap = 10
)
}
\arguments{
\item{train_mask}{a binary matrix indicating the training set}

\item{validation_mask}{a binary matrix indicating the validation set}

\item{observation}{a matrix of observed values, 0 or 1}

\item{K}{dimension of latent space}

\item{lambda}{regularization parameter}

\item{max_iter}{maximum number of iterations for collaborative filtering}

\item{tol}{tolerance for collaborative filtering, change of logit loss between iterations}

\item{cap}{limit the maximum of the absolute value of estimated logit link}
}
\value{
logistic collaborative filtering results and performance on validation set
}
\description{
training logistic collaborative filtering and test it on a validation set
}
