% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfbm.R
\name{dfbm}
\alias{dfbm}
\title{Denoise count matrices by factorization of binary masks}
\usage{
dfbm(
  count_mat,
  quantiles = seq(0.1, 0.9, 0.1),
  increment = 0.9,
  max_K = 10,
  lambdas = c(0.01, 0.1, 1),
  ncores = 1
)
}
\arguments{
\item{count_mat}{count matrix, nsamples * nfeatures}

\item{quantiles}{the candidate thresholds are selected from the quantiles of each feature values}

\item{increment}{a value between 0 and 1, at least certain percentage of entries larger than one threshold is smaller than the next threshold}

\item{max_K}{maximum number of ranks for binary matrix factorization}

\item{lambdas}{ridge penalty parameters to try for the entries in the factorized matrices}

\item{ncores}{number of cores for parallel computing, default 1}
}
\value{
a list object that includes the denoised counts, denoised probability matrices,thresholds and optimal ranks
}
\description{
Denoise count matrices by factorization of binary masks
}
