% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmfsvt.R
\name{stack_ce}
\alias{stack_ce}
\title{Stacked cross entropy over all masks}
\usage{
stack_ce(X, Y, Omega = NULL)
}
\arguments{
\item{X}{list of natural parameter matrices}

\item{Y}{list of binary matrices}

\item{Omega}{observed entry mask or NULL}
}
\value{
the summed negative log likelihood
}
\description{
Stacked cross entropy over all masks
}
\keyword{internal}
